Pablo's fixes
