--- Java/src/main/java/Planes/MyPlane.java
@@ -1,9 +1,0 @@
-package Planes;
-
-public class MyPlane  extends Plane {
-    public MyPlane(String model, int maxSpeed, int maxFlightDistance, int maxLoadCapacity) {
-        super(model, maxSpeed, maxFlightDistance, maxLoadCapacity);
-    }
-
-    //TODO implement it later
-}
