--- Java/src/main/java/Planes/experimentalPlane.java
@@ -1,41 +1,0 @@
-package Planes;
-
-import models.ClassificationLevel;
-import models.ExperimentalTypes;
-
-public class experimentalPlane extends Plane{
-
-    private ExperimentalTypes type;
-    private ClassificationLevel classificationLevel;
-
-    public experimentalPlane(String model, int maxSpeed, int maxFlightDistance, int maxLoadCapacity, ExperimentalTypes type, ClassificationLevel classificationLevel) {
-        super(model, maxSpeed, maxFlightDistance, maxLoadCapacity);
-        this.type = type;
-        this.classificationLevel = classificationLevel;
-    }
-
-    public ClassificationLevel getClassificationLevel(){
-        return classificationLevel;
-    }
-
-    public void setClassificationLevel(ClassificationLevel classificationLevel){
-        this.classificationLevel = classificationLevel;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        return super.equals(o);
-    }
-
-    @Override
-    public int hashCode() {
-        return super.hashCode();
-    }
-
-    @Override
-    public String toString() {
-        return "experimentalPlane{" +
-                "model='" + model + '\'' +
-                '}';
-    }
-}
