+++ Java/src/main/java/epam/models/ClassificationLevel.java
@@ -1,0 +1,5 @@
+package epam.models;
+
+public enum ClassificationLevel {
+	UNCLASSIFIED, SECRET, TOP_SECRET
+}
