+++ Java/src/main/java/epam/models/MilitaryType.java
@@ -1,0 +1,5 @@
+package epam.models;
+
+public enum MilitaryType {
+	FIGHTER, BOMBER, TRANSPORT
+}
