--- Java/src/main/java/models/ClassificationLevel.java
@@ -1,5 +1,0 @@
-package models;
-
-public enum ClassificationLevel {
-    UNCLASSIFIED, CONFIDENTIAL, SECRET, TOP_SECRET
-}
