--- Java/src/main/java/models/ExperimentalTypes.java
@@ -1,5 +1,0 @@
-package models;
-
-public enum ExperimentalTypes {
-    LIFTING_BODY, HYPERSONIC, HIGH_ALTITUDE, VTOL
-}
