--- Java/src/main/java/models/MilitaryType.java
@@ -1,5 +1,0 @@
-package models;
-
-public enum MilitaryType {
-    FIGHTER, BOMBER, TRANSPORT
-}
